src/requirements2.txt
accelerate==1.0.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
airportsdata==20241001
annotated-types==0.7.0
anthropic==0.37.1
antlr4-python3-runtime==4.11.1
anyio==4.6.2.post1
asttokens==3.0.0
attrs==24.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
datasets==3.0.2
decorator==5.1.1
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
docker-pycreds==0.4.0
docstring_parser==0.16
einops==0.8.0
eval_type_backport==0.2.0
executing==2.1.0
fastcore==1.7.28
filelock==3.16.1
fire==0.7.0
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.9.0
func_timeout==4.3.5
ghapi==1.0.6
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.21.0
google-auth==2.35.0
google-cloud-aiplatform==1.70.0
google-cloud-bigquery==3.26.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.5
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpc-google-iam-v1==0.13.1
grpcio==1.67.0
grpcio-status==1.67.0
guidance==0.2.0
guidance-stitch==0.1.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.1
idna==3.10
iniconfig==2.0.0
interegular==0.3.3
ipython==8.31.0
ipywidgets==8.1.5
## !! Could not determine repository location
-e /home/tarun/struct_cot/src/itergen/iter_syncode
jedi==0.19.2
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
lark==1.2.2
-e git+https://github.com/tarsur909/struct_cot.git@39f4bb8160907b9fa58b70c1a2ff99562f5e1de0#egg=latex2sympy2&subdirectory=src/math_evaluator/latex2sympy
llguidance==0.5.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
mxeval @ git+https://github.com/shubhamugare/mxeval.git@590129d063c5e7d145e8b15a83f00ff20d69a189
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.52.2
ordered-set==4.1.0
outcome==1.3.0.post0
outlines==0.1.1
outlines_core==0.1.14
packaging==24.1
pandas==2.2.3
parso==0.8.4
Pebble==5.1.0
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
ply==3.11
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycountry==24.6.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
PySocks==1.7.1
pytest==8.3.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
ray==2.40.0
referencing==0.35.1
regex==2023.8.8
requests==2.32.3
rich==13.9.3
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
selenium==4.27.1
sentencepiece==0.2.0
sentry-sdk==2.17.0
setproctitle==1.3.3
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
srsly==2.4.8
stack-data==0.6.3
sympy==1.13.1
-e git+https://github.com/shubhamugare/syncode.git@c0c7e56fe3430898494853c867f407e4917bc5f2#egg=syncode
tabulate==0.9.0
termcolor==2.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
timeout-decorator==0.5.0
together==1.3.3
tokenizers==0.19.1
torch==2.5.0
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.0
transformers-cfg==0.2.6
trio==0.28.0
trio-websocket==0.11.1
triton==3.1.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wandb==0.18.5
wcwidth==0.2.13
websocket-client==1.8.0
widgetsnbextension==4.0.13
word2number==1.1
wsproto==1.2.0
xxhash==3.5.0
yarl==1.16.0
z3-solver==4.13.4.0
